+++ src/pipeline/stack.py
@@ -123,7 +123,7 @@ class Stack:
 		Returns:
 			An iterator over the stack's items.
 		"""
-		yield from self.stack
+		yield from self._stack
 	def __repr__(self):
 		"""Return the developer-oriented representation of the stack."""
 		return f"{type(self).__name__}({self._stack!r}, maxsize={self._maxsize})"